[FAG] clean code



# message auto-generated for no-merge-commit merge:
!1057 merge clean_code into master

[FAG] clean code

Created-by: cheng_zhiwei
Commit-by: 程志炜
Merged-by: cann-robot
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：



See merge request: cann/ops-transformer!1057

+++ attention/flash_attention_score_grad/op_host/arch35/flash_attention_score_grad_tiling_s1s2_bn2gs1s2_regbase.cpp
@@ -3098,7 +3098,7 @@ void FlashAttentionScoreGradTilingUs1s2Bs2Regbase::GetParseS1S2OuterInfo(int64_t
             fBaseParams.isInvalidCol = true;
         }
         // check invalid row or col block for BN2
-        for (size_t j = 0; j < invalidS1Array.size(); j++) {
+        for (int64_t j = 0; j < static_cast<int64_t>(invalidS1Array.size()); j++) {
             if (j >= parseInfo[i][BEGIN_IDX] && j < parseInfo[i][END_IDX]) {
                 invalidS1Array[j] = true;
             }
